+++ CHANGELOG.md
@@ -3,7 +3,17 @@ Changelog
 
 [Unreleased]
 ------------
--
+### Added
+- Quiz Queue
+  - Button for hosts to add the quiz they are on to the queue.
+  - Draggable (for hosts) list of quizzes in the queue.
+  - Remove button for hosts to remove quizzes from the queue.
+  - Go to quiz button for hosts to visit the page of quizzes in the queue.
+  - Controls for hosts to toggle and customise interval for auto changing to
+    next quiz.
+    - Countdown under leaderboard displaying time left until change to next
+    quiz.
+    - Button to cancel the countdown for hosts.
 
 [v2.2.0] - 2021-08-16
 ---------------------